Fix review dates landing years in the future (string + number = concatenation)
express-validator leaves req.body values as-is by default — a number
input's value arrives as the string "1", not the number 1. Nothing
converted document_review_frequency_months before backfillReviewDates()
fed it into date.getMonth() + months: with months as a string, JS does
"7" + "1" = "71" instead of 7 + 1 = 8, and Date.setMonth(71) rolls the
date five years into the future. Reproduced the exact reported symptom
(setting "1 month" produced a 2031 review date) before fixing it.

Added .toInt() to every month-count validator that feeds this kind of
arithmetic (tenant.js, documents.js, trainings.js) and made both
addMonthsIso helpers coerce with Number() as a second line of defense,
so a string slipping through some other path can't reproduce this.

Note for the user: documents already backfilled by the buggy code have
a wrong review_date baked in — re-saving the tenant-wide frequency will
recompute them correctly now that the fix is live.

+++ src/routes/trainings.js
@@ -235,7 +235,7 @@ router.post(
   [
     body('title').trim().notEmpty().withMessage('Le titre est requis.'),
     body('type').optional({ values: 'falsy' }).trim(),
-    body('frequency_months').optional({ values: 'falsy' }).isInt({ min: 1 }).withMessage('Fréquence invalide.'),
+    body('frequency_months').optional({ values: 'falsy' }).isInt({ min: 1 }).withMessage('Fréquence invalide.').toInt(),
     body('location').optional({ values: 'falsy' }).trim().isLength({ max: 200 }),
     body('instructor').optional({ values: 'falsy' }).trim().isLength({ max: 200 }),
     body('duration').optional({ values: 'falsy' }).trim().isLength({ max: 100 }),
@@ -280,7 +280,7 @@ router.patch(
   [
     body('title').optional().trim().notEmpty().withMessage('Le titre ne peut pas être vide.'),
     body('type').optional({ values: 'falsy' }).trim(),
-    body('frequency_months').optional({ nullable: true }).isInt({ min: 1 }).withMessage('Fréquence invalide.'),
+    body('frequency_months').optional({ nullable: true }).isInt({ min: 1 }).withMessage('Fréquence invalide.').toInt(),
     body('location').optional({ nullable: true, values: 'falsy' }).trim().isLength({ max: 200 }),
     body('instructor').optional({ nullable: true, values: 'falsy' }).trim().isLength({ max: 200 }),
     body('duration').optional({ nullable: true, values: 'falsy' }).trim().isLength({ max: 100 }),
